Move monitoring radios to own modules
In subsequent PRs, these modules will get more radio-specific code, as
part of PR #3315 monitoring radio plugin work: for example, the receiving
code for each radio should move here too.

I used `git show --color-moved` to check that the moved RadioSender
definitions were not changed.
